+++ pkg/logqlmodel/stats/context_test.go
@@ -25,7 +25,7 @@ func TestResult(t *testing.T) {
 	fakeIngesterQuery(ctx)
 	fakeIngesterQuery(ctx)
 
-	res := stats.Result(2*time.Second, 2*time.Millisecond)
+	res := stats.Result(2*time.Second, 2*time.Nanosecond)
 	res.Log(util_log.Logger)
 	expected := Result{
 		Ingester: Ingester{
@@ -61,7 +61,7 @@ func TestResult(t *testing.T) {
 		},
 		Summary: Summary{
 			ExecTime:                2 * time.Second.Seconds(),
-			EnqueueTime:             2 * time.Millisecond.Seconds(),
+			QueueTime:               2 * time.Nanosecond.Nanoseconds(),
 			BytesProcessedPerSecond: int64(42),
 			LinesProcessedPerSecond: int64(50),
 			TotalBytesProcessed:     int64(84),
@@ -107,7 +107,7 @@ func TestSnapshot_JoinResults(t *testing.T) {
 		},
 		Summary: Summary{
 			ExecTime:                2 * time.Second.Seconds(),
-			EnqueueTime:             2 * time.Millisecond.Seconds(),
+			QueueTime:               2 * time.Nanosecond.Nanoseconds(),
 			BytesProcessedPerSecond: int64(42),
 			LinesProcessedPerSecond: int64(50),
 			TotalBytesProcessed:     int64(84),
@@ -116,7 +116,7 @@ func TestSnapshot_JoinResults(t *testing.T) {
 	}
 
 	JoinResults(ctx, expected)
-	res := statsCtx.Result(2*time.Second, 2*time.Millisecond)
+	res := statsCtx.Result(2*time.Second, 2*time.Nanosecond)
 	require.Equal(t, expected, res)
 }
 
@@ -179,7 +179,7 @@ func TestResult_Merge(t *testing.T) {
 		},
 		Summary: Summary{
 			ExecTime:                2 * time.Second.Seconds(),
-			EnqueueTime:             2 * time.Millisecond.Seconds(),
+			QueueTime:               2 * time.Nanosecond.Nanoseconds(),
 			BytesProcessedPerSecond: int64(42),
 			LinesProcessedPerSecond: int64(50),
 			TotalBytesProcessed:     int64(84),
@@ -226,7 +226,7 @@ func TestResult_Merge(t *testing.T) {
 		},
 		Summary: Summary{
 			ExecTime:                2 * 2 * time.Second.Seconds(),
-			EnqueueTime:             2 * 2 * time.Millisecond.Seconds(),
+			QueueTime:               2 * 2 * time.Nanosecond.Nanoseconds(),
 			BytesProcessedPerSecond: int64(42), // 2 requests at the same pace should give the same bytes/lines per sec
 			LinesProcessedPerSecond: int64(50),
 			TotalBytesProcessed:     2 * int64(84),