+++ pkg/util/marshal/legacy/marshal_test.go
@@ -80,7 +80,7 @@ var queryTests = []struct {
 				},
 				"summary": {
 					"bytesProcessedPerSecond": 0,
-					"enqueueTime": 0,
+					"queueTime": 0,
 					"execTime": 0,
 					"linesProcessedPerSecond": 0,
 					"totalBytesProcessed":0,