+++ pkg/util/marshal/marshal_test.go
@@ -86,7 +86,7 @@ var queryTests = []struct {
 					},
 					"summary": {
 						"bytesProcessedPerSecond": 0,
-						"enqueueTime": 0,
+						"queueTime": 0,
 						"execTime": 0,
 						"linesProcessedPerSecond": 0,
 						"totalBytesProcessed":0,
@@ -194,7 +194,7 @@ var queryTests = []struct {
 				},
 				"summary": {
 					"bytesProcessedPerSecond": 0,
-					"enqueueTime": 0,
+					"queueTime": 0,
 					"execTime": 0,
 					"linesProcessedPerSecond": 0,
 					"totalBytesProcessed":0,
@@ -319,7 +319,7 @@ var queryTests = []struct {
 				},
 				"summary": {
 					"bytesProcessedPerSecond": 0,
-					"enqueueTime": 0,
+					"queueTime": 0,
 					"execTime": 0,
 					"linesProcessedPerSecond": 0,
 					"totalBytesProcessed":0,